+++ docs/changelog.md
@@ -8,6 +8,7 @@
 - Improve mobile omnibar experience: ACT button and comma as separator
 - Add focus to omnibar when act button pressed
 - Remove duplicate lu command and rephrase other commands
+- Remove search bar functionality
 
 # Version 4.0.1: Post-Anki Minor Bugfixes 
 - Sync only sends diffs